Convert member model to typescript
